Remove checkboxes (anonymous contribute and public) and simplify email
Removed the checkboxes for contributing anonymously, and for make the
contribution public: the checkbox are irrevelant since the "options"
are now hard-coded -- anonymous contribution is not allowed, and all
items can be made public.
In the email, mention of a user account was removed, as well as the
login link, and the link to the archive.