Replace footer.php and call to foot() with call to partial()
Noticed that all exhibitions on dev were showing the footer information
for Burke. Why? It turns out all the exhibitions were using the
common/footer.php from the Contribution plugin instead of the one
found in the theme used by the exhibtion. Not sure why, but the
solution was to remove the footer.php from the Contribition's common subdir,
and then to use a partial in the Contribute view files instead of foot();