Allow "contributor" Omeka role to set Publication Status
On the Contributions page of the Contribution plugin, the logged-in
user can set the Publication Status, which is equivalent to making the
item public or private. Originally, this only works for users of type
"admin" or "superuser". However, we want curators to have accounts of
type "contributor".
By default, the "contributor" Omeka role does not have the permissions
to make an item public, including a user's own items. Changed the
Contribution plugin code to not check this priv. Also made it such that,
for users who are not of type/role "admin" or "superuser", these
users will only see the contributed items they own.
Reminder: the owner of a contributed item is hard-coded (CUL customization)
to be the curator of the contributed items.