Create role contributionreviewer, change access to Contribution menu
In the Contribution plugin, by default, roles superuser, admin,
contributor, and researcher have access to the Contribution menu
item in the lhs navigation bar in the admin window. However, all
'contributor' and 'researcher' accounts should not have access to
this menu; only the users who are curating contributed items should
have access to this menu. So a new role was created, contributionreviewer.
This role inherits from the contributor role, so it has all the privs that
the contributor role has. Access to the menu was given to this new role,
and access was removed for roles contributor and researcher.
Any non-admin user that needs access to the list of contributed
items should be given this role.
Note: when there are contributed items, the name of the menu item
changes from "Contribution" to "Contributed Items"