Remove Contribution information from the admin item page
The original code will list, on the item admin page,
the full name of the person who contributed (and owns) the item.
In the original code, a user was created using the email
of the contributor, as entered on the contribution form, as the
username. In the CUL-modified code, the email is ignored and
the owner of each contributed item is hard-coded to be
Elizabeth. So the Contribution information shown
in the item admin page is not correct or useful. So it will
be removed for now, and it is easy to turn it back on.