Remove calls to $acl->allow and $acl->deny involving guest role
This is part of the effort of breaking the dependency between
the Contribution plugin and the Guest User plugin.
The guest role is created by the Guest User plugin. Since we do
not want to use the guest role, the above two calls were removed.
This will not affect the CUL-modified version of the Contribution
plugin since it does not use the guest role (which would require
the creation of guest user accounts).